fix(code actions): prevent infinite loop
\o/ repro!

This took a while to find. The final repro steps are:
1. start in a journal with 2+ transactions
2. at the end of not-the-final transaction, hit enter to start a new posting line
3. this line will be empty except for 4 spaces or indentation
4. the LSP client requests code actions for this point (col 4 on blank line)
5. brrrrrr

The cause of the loop is that we need to find the transaction that contains
that point, if any, but the tree-sitter grammar was putting us into a loop
as we tried to climb the tree to see if we were in a transaction. For some
reason `node == node.parent()`, so our while loop just spun. This feels like
a bug in tree-sitter or the grammar, but it would make more sense that it's
documented and intentional for error/missing nodes and I just missed that.

Regardless, the solution is to stop trying to climb the tree if we encounter
a case where `node == node.parent()`, by comparind node ids.

This fixes the loop and hang.